Support controllers in any namespace.
Would seem, for porting reasons namely, that controllers using any
namespace should be supported in Spud.  Namespace-less and 'spud_admin'
controllers seem to work but not controllers in other namespaces.

Defaults to using the 'spud_admin' namespace if none is specified, which
preserves the current behavior.  Only difference is if any namespace is
provided it will be used instead of 'spud_admin.'